refactor: replace TheOntologyManagerView with OntologyManagerView and integrate AppHeader component
- Removed TheOntologyManagerView.vue and its associated logic.
- Created OntologyManagerView.vue with similar functionality and integrated AppHeader component.
- Added AppHeader.vue with comprehensive UI controls for ontology management.
- Updated pnpm-workspace.yaml to include only built dependencies.
