+++ frontend/src/router/index.ts
@@ -1,6 +1,6 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import { useGitHubStore } from '@/stores/github'
-import TheOntologyManagerView from '../views/TheOntologyManagerView.vue'
+import TheOntologyManagerView from "../views/OntologyManagerView.vue";
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
